requirements.txt
fastapi==0.128.0
kafka==1.3.5
kafka-python-ng
numpy==1.24.4
onnxmltools==1.14.0
onnxruntime
pandas==1.5.3
pydantic==2.12.5
python-dotenv==1.2.1
scikit_learn==1.4.1.post1
SQLAlchemy==1.4.39
xgboost==1.7.6
psycopg2-binary
redis
prefect
# Update your requirements.txt
evidently
mlflow==2.9.2
minio
boto3